Optimize memory usage for embedded systems
- Reduce TnsInfo memory footprint by decreasing windowData array size
  from 8 to 1 (MAX_TNS_WINDOWS), saving ~7.7KB per channel. TNS is
  currently only used for long blocks in faac.
- Split Huffman codeword buffer 's' into two separate arrays (data and
  len) to eliminate structure padding, saving ~4.5KB per channel on
  32/64-bit architectures.
- Use int16_t for scalefactors (sf), codebook indices (book), and
  sfb_offsets in CoderInfo, saving ~0.6KB per channel.
- Use int16_t for quantized spectrum storage in quantization and
  Huffman coding pass, saving 2KB of stack space per channel.
- Update SSE2 quantization kernel with _mm_packs_epi32 for safe 16-bit
  packing with saturation.
- Add missing <stdint.h> includes for portability of fixed-width types.
- Total persistent RAM saved: ~12.8KB per channel.
- Verified with mono and stereo encoding tests.

+++ libfaac/quantize.c
@@ -112,7 +112,7 @@ static void bmask(CoderInfo * __restrict coderInfo, faac_real * __restrict xr0, 
                   faac_real * __restrict bandenrg, faac_real * __restrict bandmaxe, int gnum, faac_real quality)
 {
   int sfb, start, end, cnt;
-  int *cb_offset = coderInfo->sfb_offset;
+  int16_t *cb_offset = coderInfo->sfb_offset;
   int last;
   faac_real avgenrg;
   faac_real powm = 0.4;
@@ -437,7 +437,7 @@ void CalcBW(unsigned *bw, int rate, SR_INFO *sr, AACQuantCfg *aacquantCfg)
 
 enum {MINSFB = 2};
 
-static void calce(faac_real * __restrict xr, const int * __restrict bands, faac_real e[NSFB_SHORT], int maxsfb,
+static void calce(faac_real * __restrict xr, const int16_t * __restrict bands, faac_real e[NSFB_SHORT], int maxsfb,
                   int maxl)
 {
     int sfb;
